pyproject.toml
[project]
name = "minecraft-server-launcher"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "shino16", email = "shinokawa.m+git@gmail.com" }
]
dependencies = [
    "discord-py>=2.4.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/minecraft_server_launcher"]
